+++ app/src/main/java/nl/ags/picum/UI/Util/RouteAdapter.java
@@ -48,7 +48,7 @@ public class RouteAdapter extends RecyclerView.Adapter<RouteAdapter.RouteViewHol
         holder.layout.setOnClickListener(new View.OnClickListener(){
             @Override
             public void onClick(View view){
-                mainActivity.showDetailsFragment(selectedRoute);
+               // mainActivity.showDetailsFragment(selectedRoute);
             }
         });
     }
