+++ app/src/main/res/layout/activity_main.xml
@@ -27,6 +27,7 @@
         android:layout_margin="16dp"
         android:backgroundTint="@color/secondary"
         android:onClick="onClickLanguageFAB"
-        />
+/>
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
