+++ app/src/main/res/values-en/strings.xml
@@ -7,4 +7,5 @@
     <string name="progress_of_route">Progres:</string>
     <string name="back">Back</string>
     <string name="Show">Show</string>
+    <string name="does_it_work">Of nee toch niet</string>
 </resources>
