+++ app/src/main/res/values-nl/strings.xml
@@ -7,4 +7,5 @@
     <string name="progress_of_route">Voortgang:</string>
     <string name="back">Terug</string>
     <string name="Show">Toon</string>
+    <string name="does_it_work">Het werkt</string>
 </resources>
