+++ app/src/main/res/values/strings.xml
@@ -6,5 +6,6 @@
     <string name="progress_of_route">Progress:</string>
     <string name="back">Back</string>
     <string name="Show">Show</string>
+    <string name="does_it_work">Does it work</string>
 
 </resources>
